Smoothing other player's movement
Other players have a velocity too, allowing to approximate their movement locally. At each reception of the player's data, the player's actual location is compared to the local approximation, and if is in the threshold, do not reset it's location (thus smooting the animation).
However, a delta between the real location and the local approximation is calculated, this delta is added (or substracted, depends of the direction) to the speed value, allowing the local approximation to close the gap between himself and the server's data over time.
Furthermore, when the pacman (so not the other players, the one behind his keyboard) change his direction, a position update is sent to the server, considerably reducing the infamous and infurating "rubberbanding" effect (see urban dictionnary).